feat: added logic to the form
