+++ src/app/components/BaseForm/helpers/consts.js
@@ -1,0 +1,14 @@
+export const initialValuesContributors = {
+  email: "",
+  phone: "",
+  message: "",
+  organization_name: "",
+  representative_name: "",
+};
+export const initialValuesVolunteer = {
+  email: "",
+  phone: "",
+  message: "",
+  name: "",
+  city: "",
+};
