+++ src/app/components/FeedbackForm/FeedbackForm.jsx
@@ -1,0 +1,22 @@
+import React from "react";
+import { useSubmitForm } from "@/app/hooks/useSubmitForm";
+import { BaseForm } from "../BaseForm/BaseForm";
+import { validationSchema } from "./helpers/validationSchema";
+import { initialValues } from "./helpers/initialValues";
+
+const FeedbackForm = ({ lng }) => {
+  const { handleSubmit, isLoading } = useSubmitForm("feedback/");
+
+  return (
+    <BaseForm
+      initialValues={initialValues}
+      validationSchema={validationSchema}
+      onSubmit={handleSubmit}
+      lng={lng}
+      isLoading={isLoading}
+      formType={"feedback"}
+    />
+  );
+};
+
+export default FeedbackForm;
