+++ src/app/components/FeedbackForm/FeedbackFormSection.jsx
@@ -1,11 +1,10 @@
 "use client";
 import React from "react";
+import { useTranslation } from "@/app/i18n/client";
 import Section from "../Section";
 import Container from "../Container";
 import Title from "../Title";
-
-import { useTranslation } from "@/app/i18n/client";
-import FeedbackForm from "./FeedbackFormMain";
+import FeedbackForm from "./FeedbackForm";
 
 export default function FeedbackFormSection({ lng }) {
   const { t } = useTranslation(lng, "title");
@@ -30,7 +29,7 @@ export default function FeedbackFormSection({ lng }) {
         xl:w-[740px] xl:h-[490px] 
         2xl:w-[852px] 2xl:h-[564px]"
           />
-          <FeedbackForm lng={lng} />
+          <FeedbackForm />
         </div>
       </Container>
     </Section>
