+++ src/app/components/FeedbackForm/helpers/initialValues.js
@@ -1,0 +1,6 @@
+export const initialValues = {
+  name: "",
+  email: "",
+  phone: "",
+  message: "",
+};
